Refactor hex-grid
