+++ packages/turf-hex-grid/.gitignore
@@ -1,0 +1,1 @@
+index.js
